requirements.txt
annotated-types==0.7.0
anyio==4.11.0
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.40.40
botocore==1.40.40
brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cryptography==46.0.1
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
django==5.2.7
et-xmlfile==2.0.0
fastapi==0.118.0
flask==3.1.2
fonttools==4.60.0
fpdf==1.7.2
fpdf2==2.8.4
git-remote-s3==0.2.5
greenlet==3.2.4
h11==0.16.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
lxml==6.0.2
markdown==3.9
markupsafe==3.0.3
matplotlib==3.10.6
narwhals==2.6.0
numpy==2.3.3
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.2
pdf2image==1.17.0
pillow==11.3.0
playwright==1.55.0
plotly==6.3.0
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pydyf==0.11.0
pyee==13.0.0
pyhanko==0.31.0
pyhanko-certvalidator==0.29.0
pyparsing==3.2.5
pypdf==6.1.1
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pyyaml==6.0.3
reportlab==4.4.4
requests==2.32.5
s3transfer==0.14.0
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
starlette==0.48.0
svglib==1.5.1
tabulate==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.37.0
weasyprint==66.0
webencodings==0.5.1
werkzeug==3.1.3
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
